requirements.txt
anthropic>=0.40.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
rich>=13.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
